GOL Splitting the projects into files
