+++ src/database.py
@@ -1,0 +1,6 @@
+# use as a DB for now
+data = {}
+# save valid JWT. {jwt: emoji}
+jwt_to_emoji = {}
+# available emojis
+emojis = {'🤢', '😍', '👽', '🥸', '🥳', '🐵'}
