+++ src/schemes.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel, Field
+
+
+class MessageScheme(BaseModel):
+    jwt: str
+    text: str = Field(max_length=255)
